requirements.txt
pyTelegramBotAPI
python-dotenv~=1.0.1
pymongo~=4.7.3